refactor(components/atom/card): api tab
